update all the types to the standard types.
note that longs in plan 9 are 32 bits, hence the change
of ulong and long to uint32_t and int32_t

+++ sys/src/9/386/random.c
@@ -101,11 +101,11 @@ randominit(void)
 /*
  *  consume random bytes from a circular buffer
  */
-ulong
-randomread(void *xp, ulong n)
+uint32_t
+randomread(void *xp, uint32_t n)
 {
-	uchar *e, *p;
-	ulong x;
+	uint8_t *e, *p;
+	uint32_t x;
 
 	p = xp;
