+++ sys/src/ape/lib/utf/runestrncat.c
@@ -26,7 +26,7 @@
 #include "utfdef.h"
 
 Rune*
-runestrncat(Rune *s1, Rune *s2, long n)
+runestrncat(Rune *s1, Rune *s2, int32_t n)
 {
 	Rune *os1;
