+++ sys/src/cmd/5i/icache.c
@@ -19,7 +19,7 @@ icacheinit(void)
 }
 
 void
-updateicache(ulong addr)
+updateicache(uint32_t addr)
 {
 	USED(addr);
 }
